requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pydantic[email]
python-multipart
sqlalchemy==2.0.25
pymysql==1.1.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
requests==2.31.0
cryptography==42.0.4
alembic==1.13.1
python-telegram-bot==20.3